+++ src/interactions/add/stats.ts
@@ -1,4 +1,4 @@
-import { type StatisticalTemplate, evalCombinaison } from "@dicelette/core";
+import { type StatisticalTemplate, evalCombination } from "@dicelette/core";
 import { ln } from "@localization";
 import { reply } from "@utils";
 import { continueCancelButtons, registerDmgButton } from "@utils/buttons";
@@ -23,7 +23,7 @@ export async function embedStatistiques(
 	const userEmbed = getEmbeds(ul, interaction.message, "user");
 	if (!userEmbed) return;
 	const statsEmbed = getEmbeds(ul, interaction.message, "stats");
-	const { combinaisonFields, stats } = getStatistiqueFields(interaction, template, ul);
+	const { combinationFields, stats } = getStatistiqueFields(interaction, template, ul);
 	//combine all embeds as one
 	userEmbed.setFooter({ text: ul("common.page", { nb: page }) });
 	//add old fields
@@ -36,9 +36,9 @@ export async function embedStatistiques(
 			inline: true,
 		});
 	}
-	const statsWithoutCombinaison = template.statistics
+	const statsWithoutCombination = template.statistics
 		? Object.keys(template.statistics)
-				.filter((stat) => !template.statistics![stat].combinaison)
+				.filter((stat) => !template.statistics![stat].combination)
 				.map((name) => name.standardize())
 		: [];
 	const embedObject = statEmbeds.toJSON();
@@ -50,17 +50,17 @@ export async function embedStatistiques(
 	}
 
 	const embedStats = Object.fromEntries(
-		Object.entries(parsedFields).filter(([key]) => statsWithoutCombinaison.includes(key))
+		Object.entries(parsedFields).filter(([key]) => statsWithoutCombination.includes(key))
 	);
-	if (Object.keys(embedStats).length === statsWithoutCombinaison.length) {
+	if (Object.keys(embedStats).length === statsWithoutCombination.length) {
 		// noinspection JSUnusedAssignment
-		let combinaison: { [name: string]: number } = {};
-		combinaison = evalCombinaison(combinaisonFields, embedStats);
-		//add combinaison to the embed
-		for (const stat of Object.keys(combinaison)) {
+		let combination: { [name: string]: number } = {};
+		combination = evalCombination(combinationFields, embedStats);
+		//add combination to the embed
+		for (const stat of Object.keys(combination)) {
 			statEmbeds.addFields({
 				name: stat.capitalize(),
-				value: `\`${combinaisonFields[stat]}\` = ${combinaison[stat]}`,
+				value: `\`${combinationFields[stat]}\` = ${combination[stat]}`,
 				inline: true,
 			});
 		}
@@ -93,18 +93,18 @@ export async function showStatistiqueModal(
 ) {
 	if (!template.statistics) return;
 	const ul = ln(interaction.locale as Djs.Locale);
-	const statsWithoutCombinaison =
+	const statsWithoutCombination =
 		Object.keys(template.statistics).filter((stat) => {
-			return !template.statistics?.[stat]?.combinaison;
+			return !template.statistics?.[stat]?.combination;
 		}) ?? [];
 	const nbOfPages =
-		Math.ceil(statsWithoutCombinaison.length / 5) >= 1
-			? Math.ceil(statsWithoutCombinaison.length / 5)
+		Math.ceil(statsWithoutCombination.length / 5) >= 1
+			? Math.ceil(statsWithoutCombination.length / 5)
 			: page;
 	const modal = new Djs.ModalBuilder()
 		.setCustomId(`page${page}`)
 		.setTitle(ul("modals.steps", { page, max: nbOfPages + 1 }));
-	let statToDisplay = statsWithoutCombinaison;
+	let statToDisplay = statsWithoutCombination;
 	if (stats && stats.length > 0) {
 		statToDisplay = statToDisplay.filter((stat) => !stats.includes(stat.unidecode()));
 		if (statToDisplay.length === 0) {
@@ -121,7 +121,7 @@ export async function showStatistiqueModal(
 	for (const stat of statsToDisplay) {
 		const cleanedName = stat.unidecode();
 		const value = statisticsLowerCase[cleanedName];
-		if (value.combinaison) continue;
+		if (value.combination) continue;
 		let msg = "";
 		if (value.min && value.max)
 			msg = ul("modals.enterValue.minAndMax", { min: value.min, max: value.max });